+++ backend/substrapp/serializers/__init__.py
@@ -6,16 +6,12 @@ from .computetask import OrchestratorTestTaskSerializer
 from .computetask import OrchestratorTrainTaskSerializer
 from .datamanager import DataManagerSerializer
 from .datasample import DataSampleSerializer
-from .datasample import OrchestratorDataSampleSerializer
-from .datasample import OrchestratorDataSampleUpdateSerializer
 from .metric import MetricSerializer
 from .metric import OrchestratorMetricSerializer
 from .model import OrchestratorModelSerializer
 
 __all__ = [
     "DataSampleSerializer",
-    "OrchestratorDataSampleSerializer",
-    "OrchestratorDataSampleUpdateSerializer",
     "DataManagerSerializer",
     "MetricSerializer",
     "OrchestratorMetricSerializer",