+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove OrchestratorDataManagerSerializer and use orchestrator client directly (#908)
 - Disabled models do not expose an address (#910)
 - Increase max page size default (#922)
+- Remove OrchestratorDataSampleSerializer and OrchestratorDataSampleUpdateSerializer and use orchestrator client directly (#911)
 
 ### Fixed
 - Compute CP dates after updating related tasks (#855)